Added settings screen
* User can change search radius to filter offers
* User recent activity limit can also be changed from settings

+++ app/src/main/java/com/pablosanchezegido/petcity/models/OfferView.kt
@@ -1,4 +1,4 @@
 package com.pablosanchezegido.petcity.models
 
 class OfferView(val id: String?, val image: String, val title: String, val startDate: Long,
-                val endDate: Long, val distance: Long, val price: Double)
+                val endDate: Long, val distance: Double, val price: Double)
